fix(runtime): complete cross-platform hardening audit
Close the 64-finding July runtime audit across memory ownership,
concurrency, I/O, archives, assets, networking, and portability while
preserving the existing language surface and zero-dependency policy.

Memory management, contexts, and collections:
- coordinate cycle collection with a managed-graph quiescence barrier, track
  reference arrays and concurrent owners, snapshot heap metadata safely, and
  make weak processing, reallocation, automatic collection, and shutdown
  finalization deterministic
- replace pool-wide hot-path locking with lifecycle epochs and class-local
  free lists, validate block ownership and double frees, and add geometric
  object-array growth plus capacity-preserving list removal
- serialize context binding, legacy-state migration, cleanup, and inherited
  child reservations; make Windows threads CRT-aware and callback boundaries
  typed; harden monitors, channels, promises, futures, and thread-pool
  construction, waits, traversal, shutdown, and trap rollback
- make Map and IntMap rehash transactional, share overflow-safe hash-table
  growth policy, distinguish Bag duplicates from OOM, and add explicit Map.Trim,
  IntMap.Trim, and channel discard APIs

I/O, archives, assets, and packaging:
- bound ZIP metadata, expansion, ratio, and extraction resources; preserve
  destination permissions during atomic replacement; clean staged output on
  returning traps; and index duplicate names during validation
- retain archive snapshots across unlocked asset loads, make first-use state
  transactional, clamp stream allocation, preserve MemStream capacity, fix
  LineReader cursor handling, and serialize watcher lifecycle and errors
- define ZPAK v2 compatibility flags and per-entry CRC-32 checksums while
  retaining v1 reads and rejecting impossible or corrupt table layouts

Network protocols and lifecycle:
- give policies, transports, clients, servers, TLS, WebSocket, SSE, SMTP,
  multipart, and async objects stable managed identity with transactional
  construction and returning-trap-safe ownership
- serialize connection-pool leases, HTTP client/session snapshots, router
  publication, server start/stop and callback bindings, retry/rate state, and
  UDP/TLS/native socket ownership across POSIX and Windows
- enforce strict HTTP, multipart, WebSocket, SSE, SMTP, TLS, and upgrade
  framing; preserve complete response identity and timeout budgets; and make
  cancellation, close, redirect, cookie, and keep-alive paths deterministic

Tooling, coverage, and Zanna Studio:
- add the gating cppcheck-runtime target, reviewed suppressions, dedicated CI
  workflow, and Windows/Unix build integration; harden sanitizer scheduling,
  stack-safety setup, source-health accounting, and installer metadata checks
- rename ZannaIDE to Zanna Studio across sources, CMake, installers, scripts,
  tests, documentation, and site pages, with one-time settings migration and
  renamed native artifacts while retaining generic ZANNA_IDE_* tooling knobs
- carry the Studio and cppcheck changes onto the pulled PowerShell Windows
  entry points, and renumber the new decision series to ADRs 0115-0136 after
  reconciling the upstream ADR namespace
- expand regression coverage for GC, ownership, contexts, collections,
  archives, ZPAK, HTTP/HTTPS, TLS, WebSocket, SSE, SMTP, async sockets,
  packaging, and the Studio migration; synchronize generated and handwritten
  runtime documentation with the resulting contracts

Validation completed before this commit request is recorded in the checked-in
audit: 45 focused CTests, 1,907 ordinary tests, the 30-test slow lane, ASan,
UBSan, TSan, runtime-surface checks, smokes, docs, platform policy, cppcheck,
source health, formatting, and diff-whitespace checks. Tests and validation
commands were not rerun during the requested pull and commit.

+++ src/runtime/network/rt_tls_server_internal.h
@@ -8,6 +8,17 @@
 // File: src/runtime/network/rt_tls_server_internal.h
 // Purpose: Internal TLS 1.3 server context + accept helpers used by
 //          HttpsServer and WssServer.
+// Key invariants:
+//   - Parsed credentials and ALPN policy are immutable after context creation.
+//   - Native descriptors remain pointer-width across the server handshake.
+//   - A valid accept call consumes its socket on success and handshake failure.
+// Ownership/Lifetime:
+//   - Callers free contexts with rt_tls_server_ctx_free after all handshakes end.
+//   - Successful accepts return one managed session reference consumed by
+//     rt_tls_close; invalid arguments do not transfer descriptor ownership.
+// Links: src/runtime/network/rt_tls.c, src/runtime/network/rt_tls.h,
+//        src/runtime/network/rt_https_server.c,
+//        src/runtime/network/rt_wss_server.c
 //
 //===----------------------------------------------------------------------===//
 #pragma once
@@ -47,8 +58,16 @@ void rt_tls_server_ctx_free(rt_tls_server_ctx_t *ctx);
 ///          handshake to completion using the cert/key + ALPN policy
 ///          in @p ctx, and returns the live session on success. Returns
 ///          NULL on any handshake failure with the cause captured via
-///          @ref rt_tls_server_last_error.
-rt_tls_session_t *rt_tls_server_accept_socket(int socket_fd, const rt_tls_server_ctx_t *ctx);
+///          @ref rt_tls_server_last_error. The pointer-width descriptor avoids
+///          narrowing WinSock `SOCKET` handles. When both arguments are valid,
+///          this function consumes and closes @p socket_fd on failure; on
+///          success the returned session owns it. An invalid context or socket
+///          returns before ownership transfer.
+/// @param socket_fd Connected native socket represented without narrowing.
+/// @param ctx Immutable parsed credential/ALPN context that outlives the session.
+/// @return Caller-owned connected session, or NULL after consuming the socket
+///         for a setup/handshake failure.
+rt_tls_session_t *rt_tls_server_accept_socket(intptr_t socket_fd, const rt_tls_server_ctx_t *ctx);
 
 /// @brief Return the most recent server-side TLS error string for diagnostics.
 const char *rt_tls_server_last_error(void);
